+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- The very first pyramid of the game now has no main-path puzzles at all; the rest of that opening journey only has one puzzle per pyramid.
 
 ## 0.26.4 - 2026-07-05
 ### Changed